+++ src/content/index.ts
@@ -1,4 +1,8 @@
-/** @module Content */
+/**
+ * @module Content
+ * @experimental
+ */
 
 export * from './api.js';
+export * from './content-scheme.js';
 export * from './procedures.js';