+++ src/file/parse.test.ts
@@ -4,7 +4,7 @@ import { ContentIdentifier } from '../cid/cid.js';
 import { Common } from '../common/common.js';
 import { createInstance } from '../instance/instance.js';
 import { FileBuilder } from './file-builder.js';
-import { parseAsFile } from './parse.js';
+import { fileScheme } from './parse.js';
 
 describe('parseAsFile', () => {
   const instance = createInstance(Common);
@@ -15,9 +15,9 @@ describe('parseAsFile', () => {
       .setMediaType('application/json')
       .setValue({ hello: 'world!' }, instance);
 
-    await expect(parseAsFile(cid, file.buffer, instance)).resolves.toEqual(file);
+    await expect(fileScheme.parse(cid, file.buffer, instance)).resolves.toEqual(file);
   });
 
   test('Throws for invalid file', () =>
-    expect(parseAsFile(cid, new Uint8Array(randomBytes(64)), instance)).rejects.toThrow());
+    expect(fileScheme.parse(cid, new Uint8Array(randomBytes(64)), instance)).rejects.toThrow());
 });